+++ core/shared/src/main/scala/com/ccadllc/cedi/dtrace/TraceContext.scala
@@ -22,10 +22,10 @@ import cats.implicits._
 import scala.language.higherKinds
 
 /**
- * Represents a cursor into the "current" [[Span]] and associated system-level metadata and is associated with an
- * effectful program `F[A]` to realize a trace over that program.
+ * Represents a cursor into the "current" [[Span]] and associated system-level information and is associated with an
+ * effectful program `F[A]` to realize a trace of that program.
  * @param currentSpan - the current [[Span]] associated with a particular effectful program.
- * @param system - a [[TraceSystem]] consisting of system-level metadata
+ * @param system - a [[TraceSystem]] consisting of system-level information
  *  which further annotates a [[Span]] when recording it along with an implementation of an
  *  [[TraceSystem#Emitter]] used to perform the recording of the span.
  * @tparam F - an effectful program type representing the program type of the [[TraceSystem]] (which in turn indicates