+++ logstash/src/test/scala/com/ccadllc/cedi/dtrace/logstash/EcsLogstashLogbackEmitterTest.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Combined Conditional Access Development, LLC.
+ * Copyright 2019 Combined Conditional Access Development, LLC.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,17 @@
  * limitations under the License.
  */
 package com.ccadllc.cedi.dtrace
+package logstash
 
-import slogging.{ SLF4JLoggerFactory, UnderlyingLoggerFactory }
+import cats.effect.IO
+import org.scalatest.WordSpec
 
-/**
- * Initializes the logging configuration for the JVM platform
- * to use the `sl4j` library.
- */
-object LoggingFactory {
-  val value: UnderlyingLoggerFactory = SLF4JLoggerFactory()
+class EcsLogstashLogbackEmitterTest extends WordSpec with TestData {
+  "EcsLogstashLogbackEmitterTest" should {
+    "work" in {
+      val system = TraceSystem(testSystemData, new EcsLogstashLogbackEmitter[IO], quarterlySalesCalculationTimer)
+      val spanRoot = Span.root[IO](quarterlySalesCalculationTimer, Span.Name("calculate-quarterly-sales")).unsafeRunSync
+      IO.unit.toTraceT.trace(TraceContext(spanRoot, system)).unsafeRunSync
+    }
+  }
 }