+++ logstash/src/test/scala/com/ccadllc/cedi/dtrace/logstash/LogstashLogbackEmitterTest.scala
@@ -22,7 +22,7 @@ import org.scalatest.WordSpec
 class LogstashLogbackEmitterTest extends WordSpec with TestData {
   "LogstashLogbackEmitterTest" should {
     "work" in {
-      val system = TraceSystem(testSystemMetadata, new LogstashLogbackEmitter[IO], quarterlySalesCalculationTimer)
+      val system = TraceSystem(testSystemData, new LogstashLogbackEmitter[IO], quarterlySalesCalculationTimer)
       val spanRoot = Span.root[IO](quarterlySalesCalculationTimer, Span.Name("calculate-quarterly-sales")).unsafeRunSync
       IO.unit.toTraceT.trace(TraceContext(spanRoot, system)).unsafeRunSync
     }